Changed Transformer and Sendable to be abstract base classes.
This decouples the rest of the code from implementation details like how the pictures are represented, etc., and lets users easily swap out different implementations. Users provide implementations for these classes, which I've demonstrated with SendableMat, which uses an OpenCV image to store the picture, and IdentityTransformer, which just sends pictures on without changing them.

My next project is to change this to to use dependency injection to allow easier swapping out of different implementations.